+++ lib/simple_jwt_manager.dart
@@ -4,6 +4,7 @@ export 'src/network/util_check_url.dart';
 export 'src/network/util_https_stub.dart'
     if (dart.library.js_interop) 'src/network/util_https_web.dart';
 export 'src/server_response/enum_server_response_status.dart';
+export 'src/server_response/enum_server_response_type.dart';
 export 'src/server_response/server_response.dart';
 export 'src/server_response/util_server_response.dart';
 export 'src/ropc_client_stub.dart'