+++ lib/src/network/util_https.dart
@@ -10,7 +10,7 @@ class UtilHttps {
   ///
   /// (ja) Httpsを構築してPOSTします。
   ///
-  /// * [url] : The URL to post to. Only https is permitted;
+  /// * [url] : The URL to post to. Only https is permitted
   /// anything else will return an error response.
   /// * [body] : The data passed in the Map will be automatically
   /// encoded according to the enum specification.
@@ -24,12 +24,22 @@ class UtilHttps {
   /// If consecutive calls are made earlier than this,
   /// they will wait until this interval before being executed.
   /// The unit is milliseconds.
+  /// * [resType] : Formatting the return value from the server.
+  ///
+  /// The return value will be formatted as follows:
+  ///
+  /// For json: ServerResponse.resBody will contain the JSON encoded return value.
+  ///
+  /// For byte: ServerResponse.resBody will contain the return value in the format { "r" : Uint8list }.
+  ///
+  /// For text: ServerResponse.resBody will contain the return value in the format { "r" : UTF-8 text }.
   static Future<ServerResponse> post(
       String url, Map<String, dynamic> body, EnumPostEncodeType type,
       {String? jwt,
       Duration timeout = const Duration(seconds: 10),
       bool adjustTiming = true,
-      intervalMs = 1200}) async {
+      intervalMs = 1200,
+      EnumServerResponseType resType = EnumServerResponseType.json}) async {
     Map<String, String> headers = {};
     if (jwt != null) {
       headers['Authorization'] = 'Bearer $jwt';
@@ -40,13 +50,15 @@ class UtilHttps {
         return customPost(url, Uri(queryParameters: body).query, headers,
             timeout: timeout,
             adjustTiming: adjustTiming,
-            intervalMs: intervalMs);
+            intervalMs: intervalMs,
+            resType: resType);
       case EnumPostEncodeType.json:
         headers['Content-Type'] = 'application/json';
         return customPost(url, jsonEncode(body), headers,
             timeout: timeout,
             adjustTiming: adjustTiming,
-            intervalMs: intervalMs);
+            intervalMs: intervalMs,
+            resType: resType);
     }
   }
 
@@ -57,7 +69,7 @@ class UtilHttps {
   /// (ja) Httpsを構築してPOSTします。
   /// これはカスタマイズ性を高めたバージョンで、簡単に利用したい場合は代わりにpostが使えます。
   ///
-  /// * [url] : The URL to post to. Only https is permitted;
+  /// * [url] : The URL to post to. Only https is permitted
   /// anything else will return an error response.
   /// * [body] : Map&lt;String, dynamic&gt;, Json encoded string, or List&lt;int&gt;.
   /// * [headers] : HTTP headers.
@@ -69,12 +81,22 @@ class UtilHttps {
   /// If consecutive calls are made earlier than this,
   /// they will wait until this interval before being executed.
   /// The unit is milliseconds.
+  /// * [resType] : Formatting the return value from the server.
+  ///
+  /// The return value will be formatted as follows:
+  ///
+  /// For json: ServerResponse.resBody will contain the JSON encoded return value.
+  ///
+  /// For byte: ServerResponse.resBody will contain the return value in the format { "r" : Uint8list }.
+  ///
+  /// For text: ServerResponse.resBody will contain the return value in the format { "r" : UTF-8 text }.
   static Future<ServerResponse> customPost(
       String url, Object? body, Map<String, String> headers,
       {Encoding? encoding,
       Duration timeout = const Duration(seconds: 10),
       bool adjustTiming = true,
-      intervalMs = 1200}) async {
+      intervalMs = 1200,
+      EnumServerResponseType resType = EnumServerResponseType.json}) async {
     try {
       if (adjustTiming) {
         await TimingManager().adjustTiming(intervalMs: intervalMs);
@@ -85,9 +107,9 @@ class UtilHttps {
               headers: headers, body: body, encoding: encoding)
           .timeout(timeout);
       if (r.statusCode == 200) {
-        return UtilServerResponse.success(r);
+        return UtilServerResponse.success(r, resType: resType);
       } else {
-        return UtilServerResponse.serverError(r);
+        return UtilServerResponse.serverError(r, resType: resType);
       }
     } on TimeoutException catch (e) {
       // response timeout