+++ lib/src/ropc_client_for_native.dart
@@ -209,7 +209,7 @@ class ROPCClientForNative {
         responseTimeout: _responseTimeout,
         adjustTiming: false);
     switch (r.resultStatus) {
-      case EnumSeverResponseStatus.success:
+      case EnumServerResponseStatus.success:
         try {
           // サーバーがトークンを返す仕様の場合は取得してログイン状態にする
           final Map<String, dynamic> tokens = jsonDecode(r.response!.body);
@@ -219,10 +219,10 @@ class ROPCClientForNative {
           // サーバーがトークンを返さない、または戻り値がJSONでは無いような場合。
           return r;
         }
-      case EnumSeverResponseStatus.timeout:
-      case EnumSeverResponseStatus.serverError:
-      case EnumSeverResponseStatus.otherError:
-      case EnumSeverResponseStatus.signInRequired:
+      case EnumServerResponseStatus.timeout:
+      case EnumServerResponseStatus.serverError:
+      case EnumServerResponseStatus.otherError:
+      case EnumServerResponseStatus.signInRequired:
         return r;
     }
   }
@@ -256,13 +256,13 @@ class ROPCClientForNative {
         responseTimeout: _responseTimeout,
         adjustTiming: false);
     switch (r.resultStatus) {
-      case EnumSeverResponseStatus.success:
+      case EnumServerResponseStatus.success:
         _clearToken();
         return r;
-      case EnumSeverResponseStatus.timeout:
-      case EnumSeverResponseStatus.serverError:
-      case EnumSeverResponseStatus.otherError:
-      case EnumSeverResponseStatus.signInRequired:
+      case EnumServerResponseStatus.timeout:
+      case EnumServerResponseStatus.serverError:
+      case EnumServerResponseStatus.otherError:
+      case EnumServerResponseStatus.signInRequired:
         return r;
     }
   }
@@ -317,7 +317,7 @@ class ROPCClientForNative {
         responseTimeout: _responseTimeout,
         adjustTiming: false);
     switch (r.resultStatus) {
-      case EnumSeverResponseStatus.success:
+      case EnumServerResponseStatus.success:
         // トークンを取得して保存
         final Map<String, dynamic> tokens = jsonDecode(r.response!.body);
         _updateJWTBuff(tokens);
@@ -327,10 +327,10 @@ class ROPCClientForNative {
               "OAuth 2.0 response error: missing token or token_type");
         }
         return r;
-      case EnumSeverResponseStatus.timeout:
-      case EnumSeverResponseStatus.serverError:
-      case EnumSeverResponseStatus.otherError:
-      case EnumSeverResponseStatus.signInRequired:
+      case EnumServerResponseStatus.timeout:
+      case EnumServerResponseStatus.serverError:
+      case EnumServerResponseStatus.otherError:
+      case EnumServerResponseStatus.signInRequired:
         return r;
     }
   }
@@ -352,12 +352,12 @@ class ROPCClientForNative {
   Future<ServerResponse> signOutAllTokens() async {
     ServerResponse res = await signOut(isRefreshToken: true);
     switch (res.resultStatus) {
-      case EnumSeverResponseStatus.success:
+      case EnumServerResponseStatus.success:
         return await signOut(isRefreshToken: false);
-      case EnumSeverResponseStatus.timeout:
-      case EnumSeverResponseStatus.serverError:
-      case EnumSeverResponseStatus.otherError:
-      case EnumSeverResponseStatus.signInRequired:
+      case EnumServerResponseStatus.timeout:
+      case EnumServerResponseStatus.serverError:
+      case EnumServerResponseStatus.otherError:
+      case EnumServerResponseStatus.signInRequired:
         return res;
     }
   }
@@ -404,7 +404,7 @@ class ROPCClientForNative {
         responseTimeout: _responseTimeout,
         adjustTiming: false);
     switch (r.resultStatus) {
-      case EnumSeverResponseStatus.success:
+      case EnumServerResponseStatus.success:
         if (isRefreshToken) {
           _refreshToken = null;
         } else {
@@ -414,10 +414,10 @@ class ROPCClientForNative {
           _tokenType = null;
         }
         return r;
-      case EnumSeverResponseStatus.timeout:
-      case EnumSeverResponseStatus.serverError:
-      case EnumSeverResponseStatus.otherError:
-      case EnumSeverResponseStatus.signInRequired:
+      case EnumServerResponseStatus.timeout:
+      case EnumServerResponseStatus.serverError:
+      case EnumServerResponseStatus.otherError:
+      case EnumServerResponseStatus.signInRequired:
         return r;
     }
   }
@@ -440,7 +440,7 @@ class ROPCClientForNative {
     }
     if (_isTokenExpired()) {
       final ServerResponse res = await refreshAndGetNewToken();
-      if (res.resultStatus != EnumSeverResponseStatus.success) {
+      if (res.resultStatus != EnumServerResponseStatus.success) {
         debugPrint(res.errorDetail);
         return null;
       }
@@ -472,7 +472,7 @@ class ROPCClientForNative {
         responseTimeout: _responseTimeout,
         adjustTiming: false);
     switch (r.resultStatus) {
-      case EnumSeverResponseStatus.success:
+      case EnumServerResponseStatus.success:
         try {
           // トークンを取得して保存
           final Map<String, dynamic> tokens = jsonDecode(r.response!.body);
@@ -486,7 +486,7 @@ class ROPCClientForNative {
         } catch (e) {
           return UtilServerResponse.otherError('Invalid token format');
         }
-      case EnumSeverResponseStatus.serverError:
+      case EnumServerResponseStatus.serverError:
         if (r.response != null) {
           // リフレッシュトークンが期限切れの場合、クライアント側のトークンをクリアする
           if (r.response!.statusCode == 401) {
@@ -495,9 +495,9 @@ class ROPCClientForNative {
           }
         }
         return r;
-      case EnumSeverResponseStatus.timeout:
-      case EnumSeverResponseStatus.otherError:
-      case EnumSeverResponseStatus.signInRequired:
+      case EnumServerResponseStatus.timeout:
+      case EnumServerResponseStatus.otherError:
+      case EnumServerResponseStatus.signInRequired:
         return r;
     }
   }