+++ lib/src/server_response/enum_server_response_type.dart
@@ -1,0 +1,1 @@
+enum EnumServerResponseType { json, byte, text }