+++ lib/src/server_response/util_server_response.dart
@@ -1,9 +1,6 @@
 import 'package:http/http.dart' as http;
+import 'package:simple_jwt_manager/simple_jwt_manager.dart';
 import 'dart:convert';
-
-import 'package:simple_jwt_manager/src/server_response/server_response.dart';
-
-import 'enum_server_response_status.dart';
 
 /// (en) A utility for formatting server responses.
 ///
@@ -12,24 +9,77 @@ class UtilServerResponse {
   /// (en) Creates a successful server response object.
   ///
   /// (ja) 成功時のサーバー応答オブジェクトを作成します。
-  static ServerResponse success(http.Response response) {
-    return ServerResponse(response, EnumSeverResponseStatus.success,
-        jsonDecode(response.body), null);
+  ///
+  /// * [response] : The server response.
+  /// * [resType] : Formatting the return value from the server.
+  ///
+  /// The return value will be formatted as follows:
+  ///
+  /// For json: ServerResponse.resBody will contain the JSON encoded return value.
+  ///
+  /// For byte: ServerResponse.resBody will contain the return value in the format { "r" : Uint8list }.
+  ///
+  /// For text: ServerResponse.resBody will contain the return value in the format { "r" : UTF-8 text }.
+  static ServerResponse success(http.Response response,
+      {EnumServerResponseType resType = EnumServerResponseType.json}) {
+    switch (resType) {
+      case EnumServerResponseType.json:
+        return ServerResponse(response, EnumServerResponseStatus.success,
+            jsonDecode(response.body), null);
+      case EnumServerResponseType.byte:
+        return ServerResponse(response, EnumServerResponseStatus.success,
+            {"r": response.bodyBytes}, null);
+      case EnumServerResponseType.text:
+        return ServerResponse(response, EnumServerResponseStatus.success,
+            {"r": response.body}, null);
+    }
   }
 
   /// (en) Creates a server response object in case of a server error.
   ///
   /// (ja) サーバーエラー時のサーバー応答オブジェクトを作成します。
-  static ServerResponse serverError(http.Response response) {
+  ///
+  /// * [response] : The server response.
+  /// * [resType] : Formatting the return value from the server.
+  ///
+  /// The return value will be formatted as follows:
+  ///
+  /// For json: ServerResponse.resBody will contain the JSON encoded return value.
+  ///
+  /// For byte: ServerResponse.resBody will contain the return value in the format { "r" : Uint8list }.
+  ///
+  /// For text: ServerResponse.resBody will contain the return value in the format { "r" : UTF-8 text }.
+  static ServerResponse serverError(http.Response response,
+      {EnumServerResponseType resType = EnumServerResponseType.json}) {
     String errorDescription = "";
     Map<String, dynamic> errorBody = {};
-    try {
-      errorBody = jsonDecode(response.body);
-      errorDescription = errorBody["error_description"];
-    } catch (e) {
-      errorDescription = "Server error. ${response.body}";
+    switch (resType) {
+      case EnumServerResponseType.json:
+        try {
+          errorBody = jsonDecode(response.body);
+          errorDescription = errorBody["error_description"] ?? "Server error.";
+        } catch (e) {
+          errorDescription = "Server error. ${response.body}";
+        }
+        break;
+      case EnumServerResponseType.byte:
+        try {
+          errorBody = {"r": response.bodyBytes};
+          errorDescription = errorBody["error_description"] ?? "Server error.";
+        } catch (e) {
+          errorDescription = "Server error. ${response.body}";
+        }
+        break;
+      case EnumServerResponseType.text:
+        try {
+          errorBody = {"r": response.body};
+          errorDescription = errorBody["error_description"] ?? "Server error.";
+        } catch (e) {
+          errorDescription = "Server error. ${response.body}";
+        }
+        break;
     }
-    return ServerResponse(response, EnumSeverResponseStatus.serverError,
+    return ServerResponse(response, EnumServerResponseStatus.serverError,
         errorBody, errorDescription);
   }
 
@@ -37,10 +87,10 @@ class UtilServerResponse {
   ///
   /// (ja) タイムアウト時のサーバー応答オブジェクトを作成します。
   ///
-  /// * [e] : The error message.
+  /// * [e] : The error object that contains the error message.
   static ServerResponse timeout(Object e) {
     return ServerResponse(
-        null, EnumSeverResponseStatus.timeout, null, e.toString());
+        null, EnumServerResponseStatus.timeout, null, e.toString());
   }
 
   /// (en) Creates a server response object when authentication is required.
@@ -48,14 +98,16 @@ class UtilServerResponse {
   /// (ja) 認証が必要になった時のサーバー応答オブジェクトを作成します。
   static ServerResponse signInRequired() {
     return ServerResponse(
-        null, EnumSeverResponseStatus.signInRequired, null, null);
+        null, EnumServerResponseStatus.signInRequired, null, null);
   }
 
   /// (en) Creates a server response object for any other error.
   ///
   /// (ja) その他のエラー発生時のサーバー応答オブジェクトを作成します。
+  ///
+  /// * [e] : The error object that contains the error message.
   static ServerResponse otherError(Object e) {
     return ServerResponse(
-        null, EnumSeverResponseStatus.otherError, null, e.toString());
+        null, EnumServerResponseStatus.otherError, null, e.toString());
   }
 }